Updated cli.js and config.js to handle showing and modifying settings store
(1) Updated yargs to a command-option structure.

(a) Updated usage information.

(b) "Countlines" command is defined as its own command now. It's options are defined within the builder callback. This command is set as the default command.

(c) Setup command, arguments, and options for the "config" command.

(2) Updated config.js to handle viewing and modifying settings store.